Expose a `VssStoreBuilder` allowing to build `VssStore` independently
Previously, `VssStore` was a private object that could only be
constructed internally via the `Builder` object. However, some users
might want to use `VssStore` independently of/before the `Node` is
constructed.

To this end, we here expose a new `VssStoreBuilder` that
allows to independently construct a `VssStore` instance that then can be
handed to `Builder::build_with_store`.

For now we keep the previous VSS-related methods on `Builder` around,
but have them reuse the `VssStoreBuilder` internally. This also allows
us to not change anything related to bindings. For bindings, we still have
to explore if/how we can expose `VssStore` in the future.