Update to new upstream `rust-lightning` with `impl Future`s
In `rust-lightning` we dropped the `Pin<Box<dyn ...>>` in favor of
`impl Future`s on most traits to make the traits native-async-ish.
Here we update to implement that change here.

However, because we avoid generics in ldk-node through concrete dyn
indirections, and having implict associated types on
`rust-lightning` traits makes them not-object-safe, we have to do
a bit of indirection. Instead, we define a `DynStoreTrait` which
re-concretizes the `rust-lightning` storage traits, allowing us to
use them `dyn`.